# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Alpha Bit Gadget Shop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: string;
  name: string;
  description: string;
  short_description: string;
  price: number;
  /** @nullable */
  sale_price?: number | null;
  category: string;
  image: string;
  stock: number;
  sku: string;
  featured?: boolean;
  on_sale?: boolean;
}

export interface ProductList {
  products: Product[];
  total: number;
  limit: number;
  skip: number;
}

export interface ProductInput {
  name: string;
  description: string;
  short_description?: string;
  price: number;
  /** @nullable */
  sale_price?: number | null;
  category: string;
  /** @nullable */
  image?: string | null;
  stock?: number;
  sku?: string;
  featured?: boolean;
  on_sale?: boolean;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  short_description?: string;
  price?: number;
  /** @nullable */
  sale_price?: number | null;
  category?: string;
  /** @nullable */
  image?: string | null;
  stock?: number;
  featured?: boolean;
  on_sale?: boolean;
}

export interface AdminProductList {
  products: Product[];
  categories: string[];
  total: number;
}

export interface Category {
  name: string;
  count: number;
}

export interface CartItem {
  product_id: string;
  quantity: number;
}

export interface CartInput {
  items: CartItem[];
  /** @nullable */
  coupon_code?: string | null;
}

export interface CartLineItem {
  product_id: string;
  name: string;
  image?: string;
  price: number;
  quantity: number;
  total: number;
  sku?: string;
}

export interface CartSummary {
  items: CartLineItem[];
  subtotal: number;
  shipping: number;
  discount?: number;
  total: number;
  free_shipping_threshold: number;
  /** @nullable */
  coupon_applied?: string | null;
}

export interface OrderInput {
  cart_items: CartItem[];
  customer_name: string;
  customer_email: string;
  customer_phone: string;
  shipping_address: string;
  shipping_city: string;
  shipping_zip: string;
  /** @nullable */
  coupon_code?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface OrderCreated {
  order_id: string;
  total: number;
  message?: string;
}

export interface Order {
  id: string;
  customer_name: string;
  customer_email: string;
  customer_phone?: string;
  shipping_address?: string;
  shipping_city?: string;
  shipping_zip?: string;
  items?: CartLineItem[];
  subtotal?: number;
  shipping?: number;
  discount?: number;
  total: number;
  status: string;
  payment_status?: string;
  /** @nullable */
  payment_id?: string | null;
  /** @nullable */
  tracking_number?: string | null;
  /** @nullable */
  notes?: string | null;
  created_at: string;
  updated_at?: string;
}

export interface AdminOrderList {
  orders: Order[];
  total: number;
}

export interface OrderStatusUpdate {
  status: string;
  /** @nullable */
  tracking_number?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface PaymentInput {
  order_id: string;
  payment_method?: string;
}

export interface PaymentResult {
  checkout_url: string;
  payment_id: string;
}

export interface PaymentStatus {
  order_id: string;
  status: string;
  /** @nullable */
  payment_id?: string | null;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  /** @nullable */
  phone?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  created_at?: string;
}

export interface AuthResult {
  token: string;
  user: User;
}

export interface ProfileUpdate {
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  current_password?: string | null;
  /** @nullable */
  new_password?: string | null;
}

export interface ForgotPasswordInput {
  email: string;
}

export interface ResetPasswordInput {
  token: string;
  new_password: string;
}

export interface CustomerSummary {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  order_count: number;
  total_spent: number;
  created_at?: string;
  tags?: string[];
}

export interface CustomerList {
  customers: CustomerSummary[];
  total: number;
}

export interface CrmStats {
  total_customers: number;
  new_this_month: number;
  top_customers: CustomerSummary[];
}

export interface Offer {
  id: string;
  code: string;
  description?: string;
  discount_type: string;
  discount_value: number;
  /** @nullable */
  min_order?: number | null;
  /** @nullable */
  max_uses?: number | null;
  uses_count?: number;
  active: boolean;
  /** @nullable */
  expires_at?: string | null;
}

export interface OfferInput {
  code: string;
  description?: string;
  discount_type: string;
  discount_value: number;
  /** @nullable */
  min_order?: number | null;
  /** @nullable */
  max_uses?: number | null;
  active?: boolean;
  /** @nullable */
  expires_at?: string | null;
}

export interface OfferUpdate {
  code?: string;
  description?: string;
  discount_type?: string;
  discount_value?: number;
  /** @nullable */
  min_order?: number | null;
  /** @nullable */
  max_uses?: number | null;
  active?: boolean;
  /** @nullable */
  expires_at?: string | null;
}

export interface OfferValidateInput {
  code: string;
  subtotal: number;
}

export interface OfferValidateResult {
  valid: boolean;
  discount_type: string;
  discount_value: number;
  discount_amount: number;
  message?: string;
}

export interface AdminStats {
  total_orders: number;
  total_revenue: number;
  total_customers: number;
  total_products: number;
  pending_orders?: number;
  paid_orders?: number;
  shipped_orders?: number;
  revenue_this_month?: number;
  orders_this_month?: number;
  recent_orders?: Order[];
}

export interface ContactInput {
  name: string;
  email: string;
  subject: string;
  message: string;
}

export interface MessageResult {
  message: string;
}

export type ListProductsParams = {
category?: string;
search?: string;
limit?: number;
skip?: number;
};

export type GetFeaturedProductsParams = {
limit?: number;
};

export type GetOfferProductsParams = {
limit?: number;
};

export type ListAdminOrdersParams = {
status?: string;
limit?: number;
skip?: number;
};

export type ListAdminCustomersParams = {
search?: string;
limit?: number;
skip?: number;
};

export type ListAdminProductsParams = {
category?: string;
search?: string;
limit?: number;
skip?: number;
};