# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Alpha Bit Gadget Shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderCreated"
        "400":
          description: Bad request
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found

  # ── PAYMENTS ──────────────────────────────────────────────────────────────
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Email already in use
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/profile:
    put:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Send password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"

  # ── CUSTOMER ──────────────────────────────────────────────────────────────
  /customer/orders:
    get:
      operationId: getCustomerOrders
      tags: [customer]
      summary: Get orders for the logged-in customer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Customer orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  # ── OFFERS / COUPONS ──────────────────────────────────────────────────────
  /offers/validate:
    post:
      operationId: validateOffer
      tags: [orders]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferValidateInput"
      responses:
        "200":
          description: Offer valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferValidateResult"
        "400":
          description: Invalid or expired

  # ── CONTACT ───────────────────────────────────────────────────────────────
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"

  # ── ADMIN ─────────────────────────────────────────────────────────────────
...
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all orders
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderList"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CartItem:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: string
        quantity:
          type: integer
...
    CartLineItem:
      type: object
      required: [product_id, name, price, quantity, total]
      properties:
        product_id:
          type: string
        name:
          type: string
        image:
          type: string
        price:
          type: number
        quantity:
          type: integer
        total:
          type: number
        sku:
          type: string
...
    OrderInput:
      type: object
      required: [cart_items, customer_name, customer_email, customer_phone, shipping_address, shipping_city, shipping_zip]
      properties:
        cart_items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        customer_name:
          type: string
        customer_email:
          type: string
        customer_phone:
          type: string
        shipping_address:
          type: string
        shipping_city:
          type: string
        shipping_zip:
          type: string
        coupon_code:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    OrderCreated:
      type: object
      required: [order_id, total]
      properties:
        order_id:
          type: string
        total:
          type: number
        message:
          type: string
    Order:
      type: object
      required: [id, customer_name, customer_email, status, total, created_at]
      properties:
        id:
          type: string
        customer_name:
          type: string
        customer_email:
          type: string
        customer_phone:
          type: string
        shipping_address:
          type: string
        shipping_city:
          type: string
        shipping_zip:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartLineItem"
        subtotal:
          type: number
        shipping:
          type: number
        discount:
          type: number
        total:
          type: number
        status:
          type: string
        payment_status:
          type: string
        payment_id:
          type: ["string", "null"]
        tracking_number:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
    AdminOrderList:
      type: object
      required: [orders, total]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        created_at:
          type: string
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        current_password:
          type: ["string", "null"]
        new_password:
          type: ["string", "null"]
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    ResetPasswordInput:
      type: object
      required: [token, new_password]
      properties:
        token:
          type: string
        new_password:
          type: string

    # ── Customer ─────────────────────────────────────────────────────────────
...
    OfferValidateInput:
      type: object
      required: [code, subtotal]
      properties:
        code:
          type: string
        subtotal:
          type: number
    OfferValidateResult:
      type: object
      required: [valid, discount_type, discount_value, discount_amount]
      properties:
        valid:
          type: boolean
        discount_type:
          type: string
        discount_value:
          type: number
        discount_amount:
          type: number
        message:
          type: string

    # ── Admin Stats ───────────────────────────────────────────────────────────
...
    ContactInput:
      type: object
      required: [name, email, subject, message]
      properties:
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
    MessageResult:
      type: object
      required: [message]
      properties:
        message:
          type: string